rearrange, add validators
