+++ modules/core/shared/src/main/scala/lucuma/core/syntax/package.scala
@@ -21,3 +21,5 @@ object all
     with ToTreeMapOps
     with ToTreeSetCompanionOps
     with ToEnumeratedOps
+    with timespan
+    with validation
